Handling Cirular Reference, Adding Support for Double, Float Cell Type and adding support for Arithmetic, logical and relational operators.

+++ src/com/orkes/cell/common/Cell.java
@@ -1,5 +1,5 @@
 package com.orkes.cell.common;
 
 public interface Cell {
-    int getValue();
+    double getValue();
 }
