+++ src/com/orkes/cell/specific/ConstantCell.java
@@ -3,14 +3,14 @@ package com.orkes.cell.specific;
 import com.orkes.cell.common.Cell;
 
 public class ConstantCell implements Cell {
-    private final int value;
+    private final double value;
 
-    public ConstantCell(int value) {
+    public ConstantCell(double value) {
         this.value = value;
     }
 
     @Override
-    public int getValue() {
+    public double getValue() {
         return value;
     }
 }
